language changes: a lot.
- `uint` and `int` are now `usize` and `isize`.
- removed some redundant uses of `[ui]size`.
- added `#[allow(unstable)]`.
- `box` is gated, temporarily switch to `Box::new`.
- slicing yields an unsized slice, added lots of `&` for that.
- macro syntax is constrained for forward compatibility,
  adapted the stateful decoder for new rules.
- use `{:?}` instead of `{}` in debugging messages.

+++ src/index/japanese/jis0208.rs
@@ -1120,7 +1120,7 @@ static FORWARD_TABLE: &'static [u16] = &[
 #[inline]
 #[stable]
 pub fn forward(code: u16) -> u32 {
-    let code = code as uint;
+    let code = code as usize;
     if code < 11104 {
         FORWARD_TABLE[code] as u32
     } else {
@@ -3309,9 +3309,9 @@ static BACKWARD_TABLE_REMAPPED: &'static [u16] = &[
 #[inline]
 #[stable]
 pub fn backward(code: u32) -> u16 {
-    let offset = (code >> 5) as uint;
-    let offset = if offset < 2048 {BACKWARD_TABLE_UPPER[offset] as uint} else {0};
-    BACKWARD_TABLE_LOWER[offset + ((code & 31) as uint)]
+    let offset = (code >> 5) as usize;
+    let offset = if offset < 2048 {BACKWARD_TABLE_UPPER[offset] as usize} else {0};
+    BACKWARD_TABLE_LOWER[offset + ((code & 31) as usize)]
 }
 
 /// Returns the index shift_jis pointer for code point `code`.
@@ -3320,7 +3320,7 @@ pub fn backward(code: u32) -> u16 {
 pub fn backward_remapped(code: u32) -> u16 {
     let value = backward(code);
     if 8272 <= value && value <= 8835 {
-        BACKWARD_TABLE_REMAPPED[(value - 8272) as uint]
+        BACKWARD_TABLE_REMAPPED[(value - 8272) as usize]
     } else {
         value
     }
