+++ src/index/singlebyte/windows_1258.rs
@@ -25,7 +25,7 @@ static FORWARD_TABLE: &'static [u16] = &[
 #[inline]
 #[stable]
 pub fn forward(code: u8) -> u16 {
-    FORWARD_TABLE[(code - 0x80) as uint]
+    FORWARD_TABLE[(code - 0x80) as usize]
 }
 
 static BACKWARD_TABLE_LOWER: &'static [u8] = &[
@@ -76,9 +76,9 @@ static BACKWARD_TABLE_UPPER: &'static [u16] = &[
 #[inline]
 #[stable]
 pub fn backward(code: u32) -> u8 {
-    let offset = (code >> 5) as uint;
-    let offset = if offset < 266 {BACKWARD_TABLE_UPPER[offset] as uint} else {0};
-    BACKWARD_TABLE_LOWER[offset + ((code & 31) as uint)]
+    let offset = (code >> 5) as usize;
+    let offset = if offset < 266 {BACKWARD_TABLE_UPPER[offset] as usize} else {0};
+    BACKWARD_TABLE_LOWER[offset + ((code & 31) as usize)]
 }
 
 #[cfg(test)]
